pyproject.toml
[project]
name = "django-template-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.35",
    "celery>=5.5.3",
    "dishka>=1.6.0",
    "django>=5.2.3",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.7.0",
    "django-prometheus>=2.3.1",
    "django-redis>=5.4.0",
    "django-storages>=1.14.6",
    "django-unfold>=0.59.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "faker>=37.4.0",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "python-decouple>=3.8",
    "redis>=6.2.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
